requirements.txt
argparse
python-dotenv
requests
pandas
matplotlib
selenium